requirements.txt
requests==2.23.0
xmltodict==0.12.0
youtube-dl==2020.09.20
bilibiliupload==0.2.3
gitpython==3.1.11
